+++ stem/descriptor/server_descriptor.py
@@ -58,6 +58,7 @@ from stem.descriptor import (
   _parse_bytes_line,
   _parse_timestamp_line,
   _parse_forty_character_hex,
+  _parse_protocol_line,
   _parse_key_block,
 )
 
@@ -96,6 +97,7 @@ SINGLE_FIELDS = (
   'protocols',
   'allow-single-hop-exits',
   'tunnelled-dir-server',
+  'proto',
   'onion-key-crosscert',
   'ntor-onion-key',
   'ntor-onion-key-crosscert',
@@ -394,6 +396,7 @@ _parse_write_history_line = functools.partial(_parse_history_line, 'write-histor
 _parse_ipv6_policy_line = _parse_simple_line('ipv6-policy', 'exit_policy_v6', func = lambda v: stem.exit_policy.MicroExitPolicy(v))
 _parse_allow_single_hop_exits_line = _parse_if_present('allow-single-hop-exits', 'allow_single_hop_exits')
 _parse_tunneled_dir_server_line = _parse_if_present('tunnelled-dir-server', 'allow_tunneled_dir_requests')
+_parse_proto_line = _parse_protocol_line('proto', 'protocols')
 _parse_caches_extra_info_line = _parse_if_present('caches-extra-info', 'extra_info_cache')
 _parse_family_line = _parse_simple_line('family', 'family', func = lambda v: set(v.split(' ')))
 _parse_eventdns_line = _parse_simple_line('eventdns', 'eventdns', func = lambda v: v == '1')
@@ -447,6 +450,7 @@ class ServerDescriptor(Descriptor):
   :var list or_addresses: **\*** alternative for our address/or_port
     attributes, each entry is a tuple of the form (address (**str**), port
     (**int**), is_ipv6 (**bool**))
+  :var stem.descriptor.ProtocolSupport protocols: supported protocols
 
   **Deprecated**, moved to extra-info descriptor...
 
@@ -463,6 +467,9 @@ class ServerDescriptor(Descriptor):
 
   .. versionchanged:: 1.5.0
      Added the allow_tunneled_dir_requests attribute.
+
+  .. versionchanged:: 1.6.0
+     Added the protocols attribute.
   """
 
   ATTRIBUTES = {
@@ -493,6 +500,7 @@ class ServerDescriptor(Descriptor):
     'hibernating': (False, _parse_hibernating_line),
     'allow_single_hop_exits': (False, _parse_allow_single_hop_exits_line),
     'allow_tunneled_dir_requests': (False, _parse_tunneled_dir_server_line),
+    'protocols': ({}, _parse_proto_line),
     'extra_info_cache': (False, _parse_caches_extra_info_line),
     'extra_info_digest': (None, _parse_extrainfo_digest_line),
     'hidden_service_dir': (None, _parse_hidden_service_dir_line),
@@ -528,6 +536,7 @@ class ServerDescriptor(Descriptor):
     'ipv6-policy': _parse_ipv6_policy_line,
     'allow-single-hop-exits': _parse_allow_single_hop_exits_line,
     'tunnelled-dir-server': _parse_tunneled_dir_server_line,
+    'proto': _parse_proto_line,
     'caches-extra-info': _parse_caches_extra_info_line,
     'family': _parse_family_line,
     'eventdns': _parse_eventdns_line,